pyproject.toml
[tool.poetry]
name = "marvin"
version = "0.1.0"
description = ""
authors = ["Joshua Yoder <me@jpfyoder.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
mcrcon = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
